+++ .gitignore
@@ -1,4 +1,5 @@
 *.class
+node_modules
 
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
