+++ RNSimInfo.ts
@@ -1,0 +1,8 @@
+/**
+*  @providesModule RNSimInfo
+*/
+'use strict'
+
+import { NativeModules } from 'react-native'
+
+export default NativeModules.RNSimInfo
