+++ android/build.gradle
@@ -7,8 +7,8 @@ android {
     defaultConfig {
         minSdkVersion 16
         targetSdkVersion 22
-        versionCode 2
-        versionName "0.1"
+        versionCode 1
+        versionName "1.0"
         ndk {
             abiFilters "armeabi-v7a", "x86"
         }
@@ -19,6 +19,8 @@ android {
 }
 
 dependencies {
+    compile fileTree(include: ['*.jar'], dir: 'libs')
+    compile "com.android.support:appcompat-v7:23.0.1"
     compile 'com.facebook.react:react-native:+'
     compile 'com.google.android.gms:play-services-gcm:+'
 }
