+++ android/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="eu.sigrlami.sim">
+
+	<uses-permission android:name="android.permission.READ_PHONE_STATE"/>
+</manifest>
